requirements.txt
# ─── Hi-C / геномика ──────────────────────────────────────
hic-straw>=1.3.1
cooler>=0.9.3
pybedtools>=0.9.1
pyranges>=0.0.129

# ─── Вычисления ───────────────────────────────────────────
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ─── Визуализация ─────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# ─── Конфигурация / утилиты ───────────────────────────────
PyYAML>=6.0.1
tqdm>=4.66.0
jinja2>=3.1.2
click>=8.1.7

# ─── Тесты ────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0